validation: system: manually cast char to int
On some systems, char is mapped to unsigned char which causes gcc to issue this warning:
test/validation/ver_abt_log_dbg/odp_system.c:23:7: error: array subscript has type ‘char’ [-Werror=char-subscripts]
Manually casting from char to int should fix issue on both (signed or unsigned) types of system

+++ test/validation/ver_abt_log_dbg/odp_system.c
@@ -19,7 +19,7 @@ static void test_odp_version_numbers(void)
 		sizeof(version_string)-1);
 
 	while (*s) {
-		if (isdigit(*s) || (strncmp(s, ".", 1) == 0)) {
+		if (isdigit((int)*s) || (strncmp(s, ".", 1) == 0)) {
 			char_ok = 1;
 			s++;
 		} else {
